+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Remove `IsTty` trait.
   Use the standard library's [`std::io::IsTerminal`](https://doc.rust-lang.org/std/io/trait.IsTerminal.html) trait instead,
   which provides equivalent functionality.
+- Remove the `derive-more` feature and `derive_more` dependency. Event variant helper
+  methods are always available.
 
 ## Changed ⚙️
 
@@ -33,7 +35,6 @@
 - Derive standard traits for "SetCursorStyle" (#909)
 - Add query_keyboard_enhancement_flags to read enabled flags (#958)
 - Add is_* and as_* methods to the event enums (#949)
-- Add a feature flag for derive_more impls (#970)
 - Update rustix to 1.0 (#982)
 
 ## Breaking ⚠️